Fix bad CI cache on macOS
The BSD Tar that ships with macOS has a bug that can cause it to create
empty tar files, which can cause bad caches to be created on GitHub
Actions.

An example of this error is here:

    https://github.com/P3KI/bendy/pull/48/checks?check_run_id=1421741031#step:9:39

And here are a couple discussions:

    https://github.com/actions/cache/issues/403
    https://github.com/actions/toolkit/pull/553

As a workaround, this patch installs GNU Tar on macos in all jobs that
use the cache. It also adds `0-` to all cache keys to make sure that
future CI builds don't hit the old, corrupted cache. This can be removed
once `Cargo.lock` changes.