Remove rustfmt required version
Unless CI is pinned to a specific version of rust, or specific nightly,
`required_version` in `rustfmt.toml` will periodically cause CI to start
failing, when a new version of rustfmt is included in rust.

I ran into this in one of my own project, and decided that removing
`required_version` was probably preferable to pinning, so that's what
this patch does.